完成motor模块的实时报告功能。一个数据中包含多个参数，如移动步数、移动方向等

+++ include/motor.h
@@ -39,7 +39,7 @@ void motor_step_backward(int step);
 
 void motor_originate();
 void motor_origin_set();
-int motor_getorigin();
+short motor_getorigin();
 
 
 #define MOTOR_SINGLE_STEP		2
